fix: close P0-03 publication seams
